user_service/pyproject.toml
[tool.poetry]
name = "online Mart"
version = "0.1.0"
description = ""
authors = ["afzalfaizi <afzaalm993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
fastapi = {extras = ["all"], version = "^0.115.3"}
sqlmodel = "^0.0.18"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
aiokafka = "^0.10.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
